liveme_django/requirements.txt
django-silk==5.0.1
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-import-export==2.9.0
django-modeltranslation==0.18.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
